Use external OSAL and CMake tools
Use the common OS abstraction layer, and the common build
tools. Rename remaining abstraction files to pnal, for P-Net Abstract
Layer. Functions and macros have not been renamed in this commit.

The common build tools are consumed as a git submodule. The p-net
repository should be cloned with the --recurse-submodules
switch. Instructions have been updated accordingly.

Already cloned repositories can be updated with:

  $ git submodule update --init --recursive
