+++ .gitmodules
@@ -1,0 +1,3 @@
+[submodule "cmake/tools"]
+	path = cmake/tools
+	url = https://github.com/rtlabs-com/cmake-tools
