+++ .travis.yml
@@ -15,8 +15,5 @@ before_install:
   - sudo apt-get -y install cmake
 
 script:
-  - mkdir build
-  - pushd build
-  - /usr/bin/cmake .. -DCMAKE_BUILD_TYPE=Debug
-  - scan-build make -j4 all check docs install
-  - popd
+  - /usr/bin/cmake -B build -S . -DCMAKE_BUILD_TYPE=Debug
+  - /usr/bin/cmake --build build --target all check docs install -j4
