+++ CMakeLists.txt
@@ -13,13 +13,18 @@
 # full license information.
 #*******************************************************************/
 
-cmake_minimum_required (VERSION 3.13)
-set(CMAKE_MODULE_PATH ${CMAKE_MODULE_PATH} "${CMAKE_SOURCE_DIR}/cmake")
+cmake_minimum_required (VERSION 3.14)
+list(APPEND CMAKE_MODULE_PATH "${CMAKE_CURRENT_SOURCE_DIR}/cmake")
+list(APPEND CMAKE_MODULE_PATH "${CMAKE_CURRENT_SOURCE_DIR}/cmake/tools")
 project (PROFINET VERSION 0.1.0)
 
-include(CTest)
+include(AddOsal)
 include(GenerateExportHeader)
 include(CMakeDependentOption)
+
+if (CMAKE_PROJECT_NAME STREQUAL PROFINET)
+  include(CTest)
+endif()
 
 # Set required standard level
 set (CMAKE_C_STANDARD 99)
@@ -153,7 +158,7 @@ configure_file (
 add_library(profinet "")
 add_executable(pn_dev "")
 
-if (BUILD_TESTING)
+if (CMAKE_PROJECT_NAME STREQUAL PROFINET AND BUILD_TESTING)
   add_executable(pf_test "")
 endif()
 
@@ -178,6 +183,8 @@ target_include_directories(profinet
   src/device
   )
 
+target_link_libraries(profinet PUBLIC osal)
+
 install (
   TARGETS profinet
   EXPORT ProfinetConfig
@@ -198,7 +205,7 @@ install (FILES
 add_subdirectory (src)
 add_subdirectory (sample_app)
 
-if (BUILD_TESTING)
+if (CMAKE_PROJECT_NAME STREQUAL PROFINET AND BUILD_TESTING)
   add_subdirectory (test)
   include(AddGoogleTest)
   add_gtest(pf_test)
@@ -207,7 +214,7 @@ endif()
 # Doxygen configuration
 cmake_policy(SET CMP0057 NEW)
 find_package(Doxygen)
-if (DOXYGEN_FOUND)
+if (CMAKE_PROJECT_NAME STREQUAL PROFINET AND DOXYGEN_FOUND)
   set(DOXYGEN_OPTIMIZE_OUTPUT_FOR_C YES)
   set(DOXYGEN_TYPEDEF_HIDES_STRUCT YES)
   set(DOXYGEN_STRIP_CODE_COMMENTS NO)
@@ -226,7 +233,7 @@ if (DOXYGEN_FOUND)
     include
     src
     )
-endif (DOXYGEN_FOUND)
+endif()
 
 include (InstallRequiredSystemLibraries)
 set (CPACK_RESOURCE_FILE_LICENSE "${PROFINET_SOURCE_DIR}/LICENSE.md")
