+++ src/common/pf_dcp.c
@@ -39,6 +39,7 @@
 #endif
 
 #include <string.h>
+#include <inttypes.h>
 
 #include "pf_includes.h"
 #include "pf_block_writer.h"
