--- src/osal/rt-kernel/assert.h
@@ -1,26 +1,0 @@
-/*********************************************************************
- *        _       _         _
- *  _ __ | |_  _ | |  __ _ | |__   ___
- * | '__|| __|(_)| | / _` || '_ \ / __|
- * | |   | |_  _ | || (_| || |_) |\__ \
- * |_|    \__|(_)|_| \__,_||_.__/ |___/
- *
- * www.rt-labs.com
- * Copyright 2018 rt-labs AB, Sweden.
- *
- * This software is dual-licensed under GPLv3 and a commercial
- * license. See the file LICENSE.md distributed with this software for
- * full license information.
- ********************************************************************/
-
-#ifndef OSAL_ASSERT
-#define OSAL_ASSERT
-
-#include_next "assert.h"
-
-static inline void assert (int result)
-{
-   ASSERT (result);
-}
-
-#endif
